+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Changes to the plugin.
 
 ### Fixed
 
+- `modulo` is now correctly identified as a numeric operator.
 - Improved quicktest error message handling to properly identify what file the error occurred in.
 
 ## [2.2.0] - 2021-06-12